summary refactor
